restructure files to use common workload properties; use logging
