+++ workloads/mnist/mnist_jax/submission.py
