+++ workloads/mnist/mnist_jax/tuning_search_space.json
