+++ workloads/mnist/mnist_pytorch/workload.py
@@ -6,6 +6,7 @@ import time
 import itertools
 from typing import Tuple
 from collections import OrderedDict
+from workloads.mnist.workload import Mnist
 
 import spec
 import torch
@@ -38,13 +39,10 @@ class _Model(nn.Module):
         return output
 
 
-class MnistWorkload(spec.Workload):
+class MnistWorkload(Mnist):
 
   def __init__(self):
     self._eval_ds = None
-
-  def has_reached_goal(self, eval_result: float) -> bool:
-    return eval_result > 0.9
 
   def _build_dataloader(self,
       data_rng: spec.RandomState,
@@ -91,28 +89,8 @@ class MnistWorkload(spec.Workload):
     """
     raise NotImplementedError
 
-  @property
-  def loss_type(self):
-    return spec.LossType.SOFTMAX_CROSS_ENTROPY
-
-  @property
-  def train_mean(self):
-    return 0.0
-
-  @property
-  def train_stddev(self):
-    return 1.0
-
   def model_params_types(self):
     pass
-
-  @property
-  def max_allowed_runtime_sec(self):
-    return 60
-
-  @property
-  def eval_period_time_sec(self):
-    return 10
 
   # Return whether or not a key in spec.ParameterTree is the output layer
   # parameters.
@@ -188,7 +166,6 @@ class MnistWorkload(spec.Workload):
     if loss_type is not spec.LossType.SOFTMAX_CROSS_ENTROPY:
       raise NotImplementedError
 
-    # TODO: resolve whether label_batch should be one_hot or not.
     return F.nll_loss(logits_batch, label_batch)
 
   def _eval_metric(self, logits, labels):
