+++ workloads/mnist/workload.py
@@ -1,0 +1,26 @@
+import spec
+
+class Mnist(spec.Workload):
+
+  def has_reached_goal(self, eval_result: float) -> bool:
+    return eval_result > 0.9
+
+  @property
+  def loss_type(self):
+    return spec.LossType.SOFTMAX_CROSS_ENTROPY
+
+  @property
+  def train_mean(self):
+    return 0.0
+
+  @property
+  def train_stddev(self):
+    return 1.0
+
+  @property
+  def max_allowed_runtime_sec(self):
+    return 60
+
+  @property
+  def eval_period_time_sec(self):
+    return 10
